STYLUMIA/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
clip-by-openai==1.0
faiss-cpu==1.7.4
opencv-python==4.8.1.78
python-jose[cryptography]==3.3.0
